+++ GurobiCodes/gurobi_helper.h
@@ -9,7 +9,7 @@
 using namespace Eigen;
 using namespace std;
 
-void define_AlphaM2SConstr(GRBModel *, vector<Matrix<float,3,3>> *, MatrixXf *, GRBVar *, GRBVar *, GRBVar *, GRBVar *, MatrixXf *,int *, int *);
+void define_AlphaM2SConstr(GRBModel *, vector<Matrix<double,3,3>> *, MatrixXd *, GRBVar *, GRBVar *, GRBVar *, GRBVar *, MatrixXd *,int *, int *);
 
 void define_CorrespondenceConstr(GRBModel *, GRBVar *, int *, int *);
 
@@ -17,18 +17,18 @@ void define_GRBVar(GRBModel *, GRBVar *, int, int, long double, long double, cha
 
 void define_phiConstr(GRBModel *, GRBVar *, GRBVar *, int *);
 
-void define_RConstr(GRBModel *, GRBVar *, MatrixXf *);
+void define_RConstr(GRBModel *, GRBVar *, MatrixXd *);
 
 void define_SOSConstr(GRBModel *, GRBVar *, GRBVar *, GRBVar *, double *, int *);
 
 void define_SOS2Constr(GRBModel *, GRBVar *);
 
-void define_TConstr(GRBModel *, GRBVar *, MatrixXf *);
+void define_TConstr(GRBModel *, GRBVar *, MatrixXd *);
 
 void define_WConstr(GRBModel *, GRBVar *);
 
 void define_3D_GRBVar(GRBModel *, GRBVar *, int, int, int, long double, long double, char, string);
 
-void provide_initialSol(GRBVar *, GRBVar *, GRBVar *, GRBVar *, GRBVar *, GRBVar *, GRBVar *, MatrixXf *, OptVariables *, int *, int *, int *);
+void provide_initialSol(GRBVar *, GRBVar *, GRBVar *, GRBVar *, GRBVar *, GRBVar *, GRBVar *, MatrixXd *, OptVariables *, int *, int *, int *);
 
 #endif
